+++ Mineral/Core/CacheBlocks.cs
@@ -5,11 +5,11 @@ namespace Mineral.Core
 {
     public class CacheChain
     {
-        public ConcurrentDictionary<int, UInt256> HeaderIndices { get; } = new ConcurrentDictionary<int, UInt256>();
+        public ConcurrentDictionary<uint, UInt256> HeaderIndices { get; } = new ConcurrentDictionary<uint, UInt256>();
         public ConcurrentDictionary<UInt256, Block> HashBlocks { get; } = new ConcurrentDictionary<UInt256, Block>();
-        private int _capacity = 40960;
+        private uint _capacity = 40960;
 
-        public int HeaderHeight => HeaderIndices.Count - 1;
+        public uint HeaderHeight => (uint)(HeaderIndices.Count - 1);
         public UInt256 HeaderHash
         {
             get
@@ -20,8 +20,8 @@ namespace Mineral.Core
             }
         }
 
-        public void SetCapacity(int capacity) { _capacity = capacity; }
-        public int GetCapacity() { return _capacity; }
+        public void SetCapacity(uint capacity) { _capacity = capacity; }
+        public uint GetCapacity() { return _capacity; }
         public ERROR_BLOCK AddBlock(Block block)
         {
             if (!HeaderIndices.TryGetValue(block.Height, out UInt256 hash))
@@ -31,11 +31,11 @@ namespace Mineral.Core
             if (!HashBlocks.TryAdd(block.Hash, block))
                 return ERROR_BLOCK.ERROR_HASH;
             if (_capacity < HashBlocks.Count)
-                HashBlocks.TryRemove(HeaderIndices[HeaderIndices.Count - _capacity - 1], out _);
+                HashBlocks.TryRemove(HeaderIndices[(uint)HeaderIndices.Count -_capacity - 1], out _);
             return ERROR_BLOCK.NO_ERROR;
         }
 
-        public Block GetBlock(int height)
+        public Block GetBlock(uint height)
         {
             if (!HeaderIndices.TryGetValue(height, out UInt256 hash))
                 return null;
@@ -48,7 +48,7 @@ namespace Mineral.Core
             return block;
         }
 
-        public bool AddHeaderIndex(int height, UInt256 hash)
+        public bool AddHeaderIndex(uint height, UInt256 hash)
         {
             return HeaderIndices.TryAdd(height, hash);
         }
